+++ tests/libpldm_firmware_update_test.cpp
@@ -1,4 +1,8 @@
 #include <endian.h>
+#include <libpldm/base.h>
+#include <libpldm/firmware_update.h>
+#include <libpldm/pldm_types.h>
+#include <libpldm/utils.h>
 
 #include <algorithm>
 #include <array>
@@ -8,11 +12,6 @@
 #include <string>
 #include <string_view>
 #include <vector>
-
-#include "libpldm/base.h"
-#include "libpldm/firmware_update.h"
-#include "pldm_types.h"
-#include "utils.h"
 
 #include <gtest/gtest.h>