+++ tests/libpldm_utils_test.cpp
@@ -1,8 +1,8 @@
+#include <libpldm/pldm_types.h>
+#include <libpldm/utils.h>
+
 #include <cstring>
 #include <vector>
-
-#include "libpldm/utils.h"
-#include "pldm_types.h"
 
 #include <gtest/gtest.h>